fix: remove cgm, maestro2 and cirrus
